server/requirements.txt
fastapi
uvicorn[standard]
python-socketio
sqlalchemy
aiosqlite
pydantic
